pyproject.toml
[project]
name = "audrey"
version = "7.0.0"
description = "Audrey AI — self-hosted multi-model LangGraph orchestrator with OpenAI-compatible API"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Bart" }]
license = { text = "Proprietary" }

dependencies = [
    # Web framework + server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",

    # HTTP client (streaming, async)
    "httpx>=0.28",

    # Orchestration
    "langgraph>=0.2.60",
    "langchain-core>=0.3.28",

    # Config + validation
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0.2",

    # Resilience
    "tenacity>=9.0",

    # Tokenization (complexity gate, context compression)
    "tiktoken>=0.8",

    # Vector DB
    "qdrant-client>=1.12",

    # Image embeddings (CLIP only; text embeddings go through Ollama)
    "sentence-transformers>=3.3",
    "pillow>=11.0",

    # Document extraction for KB ingest
    "pypdf>=5.1",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "markdown>=3.7",

    # Mime sniffing for per-user uploads (requires libmagic1 in the image)
    "python-magic>=0.4.27",

    # Filesystem watcher for KB auto-ingest
    "watchdog>=6.0",

    # Local persistence (tool-server memory + health cache)
    "aiosqlite>=0.20",

    # Misc
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
audrey = "audrey.main:run"
audrey-ingest = "audrey.kb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/audrey"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "ASYNC", "S", "PTH", "RUF"]
ignore = ["S101", "S603", "S607"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
